benchmark: move FP8 MM test and compare prequantized vLLM GEMM
