+++ benchmark/test_blas_perf_parallel.py
@@ -25,7 +25,6 @@ from typing import Generator
 
 import pytest
 import torch
-import triton
 import yaml
 
 import flag_gems
@@ -83,7 +82,6 @@ PARALLEL_RESULT_FILE_ENV = "FLAGGEMS_BENCH_RESULT_FILE"
 torch_device_object = flag_gems.runtime.backend.gen_torch_device_object()
 DEEPGEMM_N_MULTIPLE = 64
 DEEPGEMM_K_MULTIPLE = 128
-
 
 MUL_DEFAULT_SHAPES = [
     ("broadcast", (1, 1), (1, 2048)),
@@ -157,7 +155,7 @@ class BlasBenchmark(Benchmark):
         total_flops = 0
         # shape(m,k)(k,n)
         # total_flops mxnx2k
-        if self.op_name in ("mm", "mm_w8a8_fp8"):
+        if self.op_name == "mm":
             total_flops = args[0].shape[0] * args[0].shape[1] * args[1].shape[1] * 2
         # shape(m,n)(n,p)
         # total_flops mxpx(2n+1)
@@ -781,7 +779,6 @@ class ParallelBenchmarkMixin:
 
             if self.op_name in {
                 "mm",
-                "mm_w8a8_fp8",
                 "addmm",
                 "bmm",
                 "baddbmm",
@@ -802,7 +799,6 @@ class ParallelBenchmarkMixin:
 
                 if self.op_name in {
                     "mm",
-                    "mm_w8a8_fp8",
                     "bmm",
                     "w8a8_block_fp8_matmul",
                     "router_gemm",
@@ -1059,73 +1055,6 @@ class ParallelBlasBenchmark(ParallelBenchmarkMixin, BlasBenchmark):
         if Config.bench_level == BenchLevel.COMPREHENSIVE:
             return 2
         return 1
-
-
-_MM_W8A8_FP8_OUT_CACHE = {}
-_MM_W8A8_FP8_OUT_CACHE_MAX_ENTRIES = int(
-    os.environ.get("FLAGGEMS_BENCH_MM_W8A8_OUT_CACHE_MAX", "8")
-)
-
-
-def _mm_w8a8_fp8_output_dtype(a):
-    output_dtype = os.environ.get("FLAGGEMS_MM_W8A8_OUTPUT_DTYPE", "bf16").lower()
-    if output_dtype in ("bf16", "bfloat16"):
-        return torch.bfloat16
-    if output_dtype in ("fp8", "float8"):
-        fp8_e4m3 = getattr(torch, "float8_e4m3fn", None)
-        fp8_e5m2 = getattr(torch, "float8_e5m2", None)
-        if a.dtype in (fp8_e4m3, fp8_e5m2):
-            return a.dtype
-        if fp8_e4m3 is not None:
-            return fp8_e4m3
-    return torch.bfloat16
-
-
-def _mm_w8a8_fp8_out_cached(a, b):
-    out_dtype = _mm_w8a8_fp8_output_dtype(a)
-    device_index = a.device.index if a.device.index is not None else -1
-    key = (device_index, a.shape[0], b.shape[1], out_dtype)
-    out = _MM_W8A8_FP8_OUT_CACHE.get(key)
-    if out is None or out.device != a.device:
-        out = torch.empty((a.shape[0], b.shape[1]), device=a.device, dtype=out_dtype)
-        _MM_W8A8_FP8_OUT_CACHE[key] = out
-        while len(_MM_W8A8_FP8_OUT_CACHE) > _MM_W8A8_FP8_OUT_CACHE_MAX_ENTRIES:
-            _MM_W8A8_FP8_OUT_CACHE.pop(next(iter(_MM_W8A8_FP8_OUT_CACHE)))
-    else:
-        _MM_W8A8_FP8_OUT_CACHE.pop(key)
-        _MM_W8A8_FP8_OUT_CACHE[key] = out
-    return flag_gems.mm_w8a8_fp8_out(a, b, out=out)
-
-
-class ParallelMmW8A8Fp8Benchmark(ParallelBlasBenchmark):
-    SHAPE_CONFIG_KEYS = ("mm",)
-
-    def get_latency(self, op, *args, **kwargs):
-        if op is not self.torch_op:
-            # Populate prequantization, descriptor, output, and autotune caches
-            # before CUDA Graph capture so replay measures the FP8 GEMM only.
-            for _ in range(2):
-                op(*args, **kwargs)
-            torch.cuda.synchronize()
-        return triton.testing.do_bench_cudagraph(
-            lambda: op(*args, **kwargs),
-            rep=Config.repetition,
-            return_mode="median",
-        )
-
-    def set_shapes(self, shape_file_path=None):
-        super().set_shapes(shape_file_path)
-        if not shape_file_path or not os.path.isfile(shape_file_path):
-            return
-        with open(shape_file_path, "r", encoding="utf-8") as shape_file:
-            yaml_config = yaml.safe_load(shape_file) or {}
-        if "mm" not in yaml_config:
-            return
-        self.shapes = [
-            tuple(shape)
-            for shape in yaml_config["mm"].get("shapes", self.DEFAULT_SHAPES)
-        ]
-        self.shape_desc = yaml_config["mm"].get("shape_desc", self.shape_desc)
 
 
 class ParallelBaddbmmBenchmark(ParallelBenchmarkMixin, BaddbmmBenchmark):
@@ -1555,20 +1484,6 @@ def test_blas_benchmark(op_name, torch_op, input_fn, bench_cls):
         torch_op=torch_op,
         dtypes=FLOAT_DTYPES,
     )
-    bench.run()
-
-
-@pytest.mark.mm_w8a8_fp8
-def test_mm_w8a8_fp8():
-    if not hasattr(flag_gems, "mm_w8a8_fp8_out"):
-        pytest.skip("mm_w8a8_fp8 benchmark requires the Hopper W8A8 backend")
-    bench = ParallelMmW8A8Fp8Benchmark(
-        input_fn=mm_input_fn,
-        op_name="mm_w8a8_fp8",
-        torch_op=torch.Tensor.mm,
-        dtypes=FLOAT_DTYPES,
-    )
-    bench.set_gems(_mm_w8a8_fp8_out_cached)
     bench.run()
